User Account Creation
Fixed a bug where the user model was defined twice which caused another
error that made "has_secure_password" broken. Also updated the test
model file to test the user creation with several tests and all worked
accordingly. User login will be worked on at the day of this commit.