Linked Activities and Categories to Users
Altered that database so that categories and activities are now linked
to the user. With if and else statements, I can display the appropriate
activities/categories based on if the id matches for the user.

There are a couple bugs such as the "Unsorted" category will not display
with the if and else statements so I omitted it. Will try to fix later
today.