Huge Fix from Dev Merge
Merge from Dev yesterday broke a lot of my code and I fixed quite a
number of them. I believe most of the code is working but there are bugs
yet to be found and fixed.

Also have a link_to in the home view to allow a user to edit an
activity.