Password Reset Option Successful Mail
Was able to successfully email the password reset link to the correct
user. Had to setup a gmail account just for the project in order for it
to work. Remember me feature not working due to a bug and could not fix
it after many hours spent looking for possible solutions.